add database there is an issue though
Issue: event_stream is responsible for saving values in database, this means
the value only gets saved when we have a request (from html in our case) requesting
connection to event_stream. If we don't open our website homepage then the values
won't be saved in database. I had not thought about this earlier but multiple
clients opening homepage might also mess with the data (doomed).